requirements.txt
alpaca-py==0.43.2
annotated-types==0.7.0
certifi==2026.2.25
charset-normalizer==3.4.5
idna==3.11
msgpack==1.1.2
numpy==2.4.3
pandas==3.0.1
pathlib==1.0.1
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.32.5
six==1.17.0
sseclient-py==1.9.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
